refactor: create demo for original api
