--- naive-rpc-demo/demo-api/src/main/java/com/joizhang/naiverpc/demo/api/dto/User.java
@@ -1,22 +1,0 @@
-package com.joizhang.naiverpc.demo.api.dto;
-
-import lombok.*;
-
-import java.io.Serializable;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class User implements Serializable {
-
-    private String username;
-
-    private byte gender;
-
-    private short age;
-
-    private int version;
-
-}
