--- naive-rpc-demo/demo-client/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-naiverpc.application.name=naive-demo-service
-naiverpc.register.address=zookeeper://xinlin-04:2181
