+++ naive-rpc-demo/demo-original-api/demo-interface/src/main/java/com/joizhang/naiverpc/demo/dto/User.java
@@ -1,0 +1,22 @@
+package com.joizhang.naiverpc.demo.dto;
+
+import lombok.*;
+
+import java.io.Serializable;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class User implements Serializable {
+
+    private String username;
+
+    private byte gender;
+
+    private short age;
+
+    private int version;
+
+}
