+++ naive-rpc-demo/demo-original-api/demo-interface/src/main/java/com/joizhang/naiverpc/demo/service/HelloService.java
@@ -1,0 +1,7 @@
+package com.joizhang.naiverpc.demo.service;
+
+public interface HelloService {
+
+    String hello(String name);
+
+}
