+++ naive-rpc-demo/demo-original-api/demo-interface/src/main/java/com/joizhang/naiverpc/demo/service/UserService.java
@@ -1,0 +1,9 @@
+package com.joizhang.naiverpc.demo.service;
+
+import com.joizhang.naiverpc.demo.dto.User;
+
+public interface UserService {
+
+    User incrementVersion(User user);
+
+}
