+++ naive-rpc-demo/demo-original-api/demo-provider/src/main/java/com/joizhang/naiverpc/demo/provider/service/HelloServiceImpl.java
@@ -1,0 +1,16 @@
+package com.joizhang.naiverpc.demo.provider.service;
+
+import com.joizhang.naiverpc.demo.service.HelloService;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class HelloServiceImpl implements HelloService {
+
+    @Override
+    public String hello(String name) {
+        String msg = "Hello, " + name;
+        log.info(msg);
+        return msg;
+    }
+
+}
