+++ naive-rpc-demo/demo-original-api/demo-provider/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+naiverpc.application.name=demo-service
+naiverpc.register.address=zookeeper://xinlin-04:2181
