+++ naive-rpc-demo/demo-original-api/pom.xml
@@ -1,0 +1,26 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.joizhang</groupId>
+        <artifactId>naive-rpc-demo</artifactId>
+        <version>2021.07</version>
+    </parent>
+
+    <artifactId>demo-original-api</artifactId>
+    <packaging>pom</packaging>
+
+    <name>demo-original-api</name>
+    <modules>
+        <module>demo-interface</module>
+        <module>demo-provider</module>
+        <module>demo-consumer</module>
+    </modules>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
+    <dependencies>
+    </dependencies>
+</project>
