--- naive-rpc-demo/demo-service/src/main/java/com/joizhang/naiverpc/demo/service/HelloServiceImpl.java
@@ -1,16 +1,0 @@
-package com.joizhang.naiverpc.demo.service;
-
-import com.joizhang.naiverpc.demo.api.service.HelloService;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class HelloServiceImpl implements HelloService {
-
-    @Override
-    public String hello(String name) {
-        String msg = "Hello, " + name;
-        log.info(msg);
-        return msg;
-    }
-
-}
